HW4/HW4_Falcon/requirements.txt
falcon==3.1.0
numpy==1.23.1
opencv_python_headless==4.6.0.66
Pillow==9.2.0
requests==2.28.1
